Allows modal to open without a trigger ref
Makes the `triggerRef` prop for the Modal component optional.

If a `triggerRef` is not provided, the modal's animation origin will default to the center of the screen. This enables the Modal component to be used in scenarios where there isn't a specific UI element to animate from, such as the `ReloadWarning` component.